+++ test/server/routes/DefaultRoute.ts
@@ -1,10 +1,12 @@
 ///////////////////////////////////////////////////////////////////////////////
 // Copyright (C) 2018 AcceleratXR, Inc. All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
-import { Route, Get, User, Init, Auth, WebSocket, Socket } from "../../../src/decorators/RouteDecorators";
-import { Logger } from "@composer-js/core";
+import { Route, Get, User, Auth, WebSocket, Socket, Query } from "../../../src/decorators/RouteDecorators";
+import { ApiError, Logger } from "@composer-js/core";
+import { Init } from "../../../src/decorators/ObjectDecorators";
 import * as ws from "ws";
 import { Description, Returns } from "../../../src/decorators/DocDecorators";
+import { ApiErrors, ApiErrorMessages } from "../../../src/ApiErrors";
 
 const logger = Logger();
 
@@ -14,12 +16,12 @@ class DefaultRoute {
     /**
      * Initializes a new instance with the specified defaults.
      */
-    constructor() { 
+    constructor() {
         // NO-OP
     }
 
     @Init
-    private async initialize() { 
+    private async initialize() {
         // NO-OP
     }
 
@@ -42,9 +44,7 @@ class DefaultRoute {
     @Description("Throws a 400-level error and returns the error as the response body.")
     @Returns([null])
     protected async throwError(): Promise<any> {
-        let err: any = new Error("This is a test.");
-        err.status = 400;
-        throw err;
+        throw new ApiError(ApiErrors.INVALID_REQUEST, 400, "This is a test.");
     }
 
     @WebSocket("connect")
@@ -66,10 +66,17 @@ class DefaultRoute {
             });
             ws.send(`hello ${user.uid}`);
         } else {
-            const error: any = new Error("No user authenticated.");
-            error.status = 401;
-            throw error;
+            throw new ApiError(ApiErrors.AUTH_REQUIRED, 401, ApiErrorMessages.AUTH_REQUIRED);
         }
+    }
+
+    @WebSocket("connect-query")
+    @Description("Establishes a socket connection that responds to all messages with the query message and message `echo ${message} ${msg}` or `echo ${msg}`.")
+    protected wsConnectQuery(@Query("message") message, @Socket ws: ws, @User user?: any): void {
+        ws.on("message", (msg) => {
+            ws.send(`echo ${message ? `${message} ` : ''}${msg}`);
+        });
+        ws.send(`hello ${user && user.uid ? user.uid : "guest"}`);
     }
 }