+++ RELEASE_NOTES.rst
@@ -2,6 +2,15 @@
 Release Notes
 =============
 
+v4.1.0
+======
+
+- Upgraded dependencies
+- Updated repo documents
+- Removed deprecated RouteDecorator Init
+- Updated to allow WebSocket Request to pass query arguments
+- Leverage new ApiError class for error reporting
+
 v4.0.0
 ======